requirements.txt
# Image Dataset Management System - Python Dependencies

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
pydantic>=2.5.0

# HTTP Client (for image proxying)
httpx>=0.25.0

# Optional: Database Support (for Shop module)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Optional: Image Processing
Pillow>=10.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
